refactor: Extract setupAppContext from browser/intents entrypoints
Same code, so it makes sense to extract it
