+++ src/targets/intents/index.jsx
@@ -4,71 +4,25 @@ import { render } from 'react-dom'
 import { HashRouter } from 'react-router-dom'
 import 'cozy-ui/transpiled/react/stylesheet.css'
 import 'cozy-ui/dist/cozy-ui.min.css'
-import MostRecentCozyClient, {
-  CozyProvider as MostRecentCozyClientProvider
-} from 'cozy-client'
+import { CozyProvider } from 'cozy-client'
 import { I18n } from 'cozy-ui/react/I18n'
 
-import configureStore from 'store/configureStore'
-import { CozyClient, CozyProvider } from 'lib/redux-cozy-client'
-
-import { Application } from 'cozy-doctypes'
+import { CozyProvider as LegacyCozyProvider } from 'lib/redux-cozy-client'
 
 import IntentHandler from 'containers/IntentHandler'
+import { setupAppContext } from '../../appContext'
 
 import 'styles/intents.styl'
 
-const lang = document.documentElement.getAttribute('lang') || 'en'
-const context = window.context || 'cozy'
-
-const ACCOUNTS_DOCTYPE = 'io.cozy.accounts'
-
 document.addEventListener('DOMContentLoaded', () => {
-  const root = document.querySelector('[role=application]')
-  const appData = root.dataset
-
-  const legacyClient = new CozyClient({
-    cozyURL: `//${appData.cozyDomain}`,
-    token: appData.cozyToken
-  })
-
-  // New improvements must be done with CozyClient
-  const cozyClient = new MostRecentCozyClient({
-    uri: `${window.location.protocol}//${appData.cozyDomain}`,
-    schema: {
-      app: Application.schema,
-      accounts: {
-        doctype: ACCOUNTS_DOCTYPE,
-        attributes: {},
-        relationships: {
-          master: {
-            type: 'has-one',
-            doctype: ACCOUNTS_DOCTYPE
-          }
-        }
-      },
-      permissions: {
-        doctype: 'io.cozy.permissions',
-        attributes: {}
-      },
-      triggers: {
-        doctype: 'io.cozy.triggers'
-      }
-    },
-    token: appData.cozyToken
-  })
-
-  // store
-  const store = configureStore(legacyClient, cozyClient, context, {
-    lang
-  })
+  const { client, data, store, lang, context } = setupAppContext()
 
   render(
-    <MostRecentCozyClientProvider client={cozyClient}>
-      <CozyProvider
-        domain={appData.cozyDomain}
+    <CozyProvider client={client}>
+      <LegacyCozyProvider
+        domain={data.cozyDomain}
         store={store}
-        client={cozyClient}
+        client={client}
         secure={!__DEVELOPMENT__}
       >
         <I18n
@@ -77,11 +31,11 @@ document.addEventListener('DOMContentLoaded', () => {
           context={context}
         >
           <HashRouter>
-            <IntentHandler appData={appData} />
+            <IntentHandler appData={data} />
           </HashRouter>
         </I18n>
-      </CozyProvider>
-    </MostRecentCozyClientProvider>,
+      </LegacyCozyProvider>
+    </CozyProvider>,
     document.querySelector('[role=application]')
   )
 })
